Fix spinner behavior in MultiSwap module
